requirements.txt
Django==1.5.1
PIL==1.1.7
PyPDF2==1.19
amqp==1.3.3
anyjson==0.3.3
billiard==3.3.0.13
boto==2.20.1
celery==3.1.7
dj-database-url==0.2.1
dj-static==0.0.5
django-appconf==0.6
django-celery==3.1.1
django-celery-email==1.0.4
django-compressor==1.3
django-extensions==1.2.5
django-storages==1.1.8
django-widget-tweaks==1.3
djangorestframework==2.3.10
djrill==0.7.0
git-remote-helpers==0.1.0
gunicorn==17.5
kombu==3.0.8
psycopg2==2.5.1
pytz==2013.8
requests==2.1.0
six==1.4.1
static==0.4
wsgiref==0.1.2
